+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # pyubx2 Release Notes
 
+### RELEASE 1.0.15
+
+ENHANCEMENTS:
+
+1. The ubxdump.py example has been moved into the main pyubx2 module and configured as a setup entry point. It is now available as a 
+simple command line utility. See [Command Line Utility](https://github.com/semuconsulting/pyubx2/blob/master/README.md#cli) for usage.
+
 ### RELEASE 1.0.14
 
 ENHANCEMENTS: